requirement.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.4
action-tutorials-py==0.20.4
actionlib-msgs==4.2.4
ament-cmake-test==1.3.9
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-types==0.7.0
asttokens==2.4.1
blis==0.7.11
bond==3.0.2
Brotli @ file:///croot/brotli-split_1714483155106/work
builtin-interfaces==1.2.1
catalogue==2.0.10
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpathlib==0.19.0
composition-interfaces==1.2.1
confection==0.1.5
contourpy==1.3.0
cv-bridge==3.2.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
demo-nodes-py==0.20.4
diagnostic-msgs==4.2.4
diagnostic-updater==3.1.2
dill==0.3.8
docker-pycreds==0.4.0
domain-coordinator==0.10.0
dwb-msgs==1.1.15
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
exceptiongroup==1.2.2
executing==2.1.0
fastai==2.7.11
fastcore==1.5.55
fastdownload==0.0.7
fastprogress==1.0.3
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.53.1
fsspec==2024.9.0
geographic-msgs==1.0.6
geometry-msgs==4.2.4
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
huggingface-hub==0.24.6
idna @ file:///croot/idna_1714398848350/work
image-geometry==3.2.1
importlib_resources==6.4.4
interactive-markers==2.3.2
ipython==8.18.1
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.0
language_data==1.2.0
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.4
map-msgs==2.1.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
message-filters==4.3.4
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mpmath @ file:///croot/mpmath_1690848262763/work
murmurhash==1.0.10
nav-2d-msgs==1.1.15
nav-msgs==4.2.4
nav2-common==1.1.15
nav2-msgs==1.1.15
nav2-simple-commander==1.0.0
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
osrf-pycommon==2.0.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pcl-msgs==1.0.0
pendulum-msgs==0.20.4
pexpect==4.9.0
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.2.2
plotly==5.24.0
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.4
rcl-interfaces==1.2.1
rclpy==3.3.13
rcutils==5.1.6
requests @ file:///croot/requests_1721410876868/work
resource-retriever==3.1.2
rich==13.8.0
rmw-dds-common==1.6.0
robot-localization==3.5.3
ros2action==0.18.10
ros2bag==0.15.11
ros2cli==0.18.10
ros2component==0.18.10
ros2doctor==0.18.10
ros2interface==0.18.10
ros2launch==0.19.7
ros2lifecycle==0.18.10
ros2multicast==0.18.10
ros2node==0.18.10
ros2param==0.18.10
ros2pkg==0.18.10
ros2run==0.18.10
ros2service==0.18.10
ros2topic==0.18.10
rosbag2-interfaces==0.15.11
rosbag2-py==0.15.11
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
sentry-sdk==2.13.0
setproctitle==1.3.3
shape-msgs==4.2.4
shellingham==1.5.4
six==1.16.0
slam-toolbox==2.6.8
smart-open==7.0.4
smclib==3.0.2
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sros2==0.10.5
srsly==2.4.8
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy @ file:///croot/sympy_1724938189289/work
teleop-twist-keyboard==2.4.0
tenacity==9.0.0
tf2-geometry-msgs==0.25.7
tf2-kdl==0.25.7
tf2-msgs==0.25.7
tf2-py==0.25.7
tf2-ros-py==0.25.7
tf2-tools==0.25.7
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.0
topic-monitor==0.20.4
torch==1.13.1
torchaudio==0.13.1
torchdata==0.5.1
torchtext==0.14.1
torchvision==0.14.1
tqdm==4.66.5
traitlets==5.14.3
trajectory-msgs==4.2.4
triton==3.0.0
turtlesim==1.4.2
typer==0.12.5
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
ultralytics==8.2.88
ultralytics-thop==2.0.6
unique-identifier-msgs==2.2.1
urllib3 @ file:///croot/urllib3_1718912636303/work
visualization-msgs==4.2.4
wandb==0.17.9
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0
xacro==2.0.8
zipp==3.20.1
